# backend/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Quotes-Parts - Authentification",
        "description": "Documentation des routes d’authentification (login, logout, me).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/utilisateurs": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Lister tous les utilisateurs",
                "description": "Retourne la liste complète des utilisateurs enregistrés.",
                "operationId": "27a03cbccef26b92bf6d80712f8b8ab8",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Utilisateur"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Créer un nouvel utilisateur",
                "description": "Ajoute un nouvel utilisateur dans le système.",
                "operationId": "dded82ccae378b45b0c84f76ec078823",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "UTI_NOM",
                                    "UTI_PRENOM",
                                    "UTI_USERNAME",
                                    "UTI_PASSWORD",
                                    "GRP_CODE"
                                ],
                                "properties": {
                                    "UTI_NOM": {
                                        "type": "string"
                                    },
                                    "UTI_PRENOM": {
                                        "type": "string"
                                    },
                                    "UTI_USERNAME": {
                                        "type": "string"
                                    },
                                    "UTI_PASSWORD": {
                                        "type": "string"
                                    },
                                    "UTI_SEXE": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "UTI_AVATAR": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "GRP_CODE": {
                                        "type": "string"
                                    },
                                    "REG_CODE": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès"
                    },
                    "409": {
                        "description": "Utilisateur déjà existant"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/utilisateurs/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Afficher les détails d’un utilisateur",
                "description": "Retourne les informations détaillées d’un utilisateur spécifique.",
                "operationId": "565dd7bc723d9a51b807156bd330fb2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant unique de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Utilisateur"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Mettre à jour un utilisateur",
                "description": "Modifie les informations d’un utilisateur existant.",
                "operationId": "08fd8a9313115a94181f3651f05b6fa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "UTI_NOM": {
                                        "type": "string"
                                    },
                                    "UTI_PRENOM": {
                                        "type": "string"
                                    },
                                    "UTI_SEXE": {
                                        "type": "string"
                                    },
                                    "UTI_AVATAR": {
                                        "type": "string"
                                    },
                                    "UTI_USERNAME": {
                                        "type": "string"
                                    },
                                    "UTI_PASSWORD": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "GRP_CODE": {
                                        "type": "string"
                                    },
                                    "REG_CODE": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour avec succès"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "409": {
                        "description": "Conflit : utilisateur déjà existant"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Supprimer un utilisateur",
                "description": "Supprime un utilisateur du système par son code.",
                "operationId": "c0fc30a3d234d962f29c73d25304de83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé avec succès"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "403": {
                        "description": "Impossible de supprimer un administrateur"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/utilisateurs/{id}/toggle-status": {
            "patch": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Changer le statut d’un utilisateur (activer/désactiver)",
                "description": "Bascule le statut d’un utilisateur : actif ↔ inactif.",
                "operationId": "b9439d513c9a75c4dc997283fdf75cbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut changé avec succès"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "403": {
                        "description": "Impossible de désactiver un administrateur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Utilisateur": {
                "properties": {
                    "UTI_CODE": {
                        "description": "Code unique de l'utilisateur",
                        "type": "string"
                    },
                    "UTI_NOM": {
                        "description": "Nom de l'utilisateur",
                        "type": "string"
                    },
                    "UTI_PRENOM": {
                        "description": "Prénom de l'utilisateur",
                        "type": "string"
                    },
                    "UTI_SEXE": {
                        "description": "Sexe de l'utilisateur",
                        "type": "string",
                        "nullable": true
                    },
                    "UTI_USERNAME": {
                        "description": "Nom d'utilisateur",
                        "type": "string"
                    },
                    "UTI_AVATAR": {
                        "description": "Avatar de l'utilisateur",
                        "type": "string",
                        "nullable": true
                    },
                    "UTI_STATUT": {
                        "description": "Statut de l'utilisateur (1=actif, 0=inactif)",
                        "type": "integer"
                    },
                    "GRP_CODE": {
                        "description": "Code du groupe",
                        "type": "string"
                    },
                    "REG_CODE": {
                        "description": "Code de la régie",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}